+++ src/components/campaigns/index.ts
@@ -1,3 +1,4 @@
+export { CampaignEditDialog } from './CampaignEditDialog';
 export { CampaignsOverview } from './CampaignsOverview';
 export { CampaignsList } from './CampaignsList';
 export { CampaignDetails } from './CampaignDetails';